Added draft code into stdconfig.
git-svn-id: svn://pmwiki.org/pmwiki@1178 524c5546-5005-0410-9a3e-e25e191bd360

+++ trunk/scripts/draft.php
@@ -15,13 +15,14 @@ if ($action == 'edit')
     'accesskey' => XL('ak_savedraft')));
 
 function EditDraft(&$pagename, &$page, &$new) {
-  global $WikiDir, $DraftSuffix;
+  global $WikiDir, $DraftSuffix, $DeleteKeyPattern;
+  SDV($DeleteKeyPattern, "^\\s*delete\\s*$");
   SDV($DraftSuffix, '-Draft');
   $basename = preg_replace("/$DraftSuffix\$/", '', $pagename);
   $draftname = $basename . $DraftSuffix;
   if ($_POST['postdraft']) 
     { $pagename = $draftname; return; }
-  if ($_POST['post']) { 
+  if ($_POST['post'] && !preg_match("/$DeleteKeyPattern/", $new['text'])) { 
     $pagename = $basename; 
     $page = ReadPage($basename);
     $WikiDir->delete($draftname);
